Some code clean up and rearranging. Fixed error where I forgot to reduce before adding to heap.
